requirements.txt
google-genai>=0.5.0
pytrends>=4.9.2
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.20.0
python-dotenv>=1.0.0
requests>=2.31.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pydantic>=2.0.0
feedparser>=6.0.0
scikit-learn>=1.0.0
